requirements.txt
mcp
python-dateutil
aiohttp
streamlit
plotly
pandas
slack-sdk
python-dotenv